requirements.txt
streamlit
python-docx
comtypes; platform_system == "Windows"
pythoncom; platform_system == "Windows"
